requirements.txt
requests
feedparser
tqdm
PyMuPDF            # fitz
pdfminer.six
beautifulsoup4
sentence-transformers
faiss-cpu
transformers
datasets
scikit-learn
fastapi
uvicorn
pandas
label-studio-sdk   # optional for annotation
